+++ spot_ros2_control/examples/src/wiggle_arm.cpp
@@ -11,7 +11,7 @@
 
 #include "rclcpp/rclcpp.hpp"
 #include "sensor_msgs/msg/joint_state.hpp"
-#include "spot_ros2_control/spot_ros2_control_utilities.hpp"
+#include "spot_ros2_control/spot_joint_map.hpp"
 #include "std_msgs/msg/float64_multi_array.hpp"
 
 enum class WiggleState { WIGGLE_DOWN, WIGGLE_MIDDLE, WIGGLE_UP, RESET };
