+++ spot_ros2_control/hardware/include/spot_ros2_control/spot_hardware_interface.hpp
@@ -31,7 +31,7 @@
 #include "rclcpp/macros.hpp"
 #include "rclcpp_lifecycle/node_interfaces/lifecycle_node_interface.hpp"
 #include "rclcpp_lifecycle/state.hpp"
-#include "spot_ros2_control/spot_ros2_control_utilities.hpp"
+#include "spot_ros2_control/spot_joint_map.hpp"
 #include "spot_ros2_control/visibility_control.h"
 
 #include "bosdyn/client/lease/lease_keepalive.h"
